Rework project class layout and simplify imports - wip
