+++ .gitignore
@@ -1,10 +1,3 @@
-/RUNNING_PID
-/logs/
-/project/*-shim.sbt
-/project/project/
-/project/target/
-/target/
-
 *.class
 *.sbtserver*
 *.log
@@ -12,11 +5,11 @@
 # sbt specific
 .cache/
 .history/
-.idea/
 .lib/
 project/.sbtboot
-
-.idea_modules/
+project/metals.sbt
+target/
+*.worksheet.sc
 
 dist/*
 *.class
