+++ src/main/scala/com/dataintuitive/luciuscore/filters/package.scala
@@ -1,0 +1,7 @@
+package com.dataintuitive.luciuscore
+
+package object filters extends FiltersTrait {
+
+  type Filters = List[Filter]
+
+}
