+++ src/main/scala/com/dataintuitive/luciuscore/model/package.scala
@@ -1,0 +1,3 @@
+package com.dataintuitive.luciuscore
+
+package object model extends model.v4.ModelTrait
