+++ src/main/scala/com/dataintuitive/luciuscore/model/v3/lenses/SampleAnnotationsLenses.scala
@@ -1,7 +1,8 @@
-package com.dataintuitive.luciuscore.lenses
+package com.dataintuitive.luciuscore
+package model.v3
+package model.v3.lenses
 
 import scalaz.Lens
-import com.dataintuitive.luciuscore.model.v3.Model._
 
 object SampleAnnotationsLenses extends Serializable {
