+++ src/main/scala/com/dataintuitive/luciuscore/model/v3/package.scala
@@ -1,0 +1,4 @@
+package com.dataintuitive.luciuscore
+package model
+
+package object v3 extends ModelTrait
