+++ src/main/scala/com/dataintuitive/luciuscore/model/v4/lenses/PerturbationLenses.scala
@@ -1,7 +1,8 @@
-package com.dataintuitive.luciuscore.model.v4.lenses
+package com.dataintuitive.luciuscore
+package model.v4
+package model.v4.lenses
 
 import scalaz.Lens
-import com.dataintuitive.luciuscore.model.v4.Model._
 
 object PerturbationLenses extends Serializable {
