+++ src/main/scala/com/dataintuitive/luciuscore/model/v4/package.scala
@@ -1,0 +1,3 @@
+package com.dataintuitive.luciuscore.model
+
+package object v4 extends ModelTrait
