+++ src/test/scala/com/dataintuitive/luciuscore/io/IOTest.scala
@@ -1,12 +1,16 @@
-package com.dataintuitive.luciuscore.io
+package com.dataintuitive.luciuscore
+package io
+
+import model.v3._
 
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import com.dataintuitive.test.BaseSparkContextSpec
-import com.dataintuitive.luciuscore.io.StatsIO._
-import com.dataintuitive.luciuscore.io.SampleCompoundRelationsIO._
-import com.dataintuitive.luciuscore.io.CompoundAnnotationsIO._
-import com.dataintuitive.luciuscore.io.RanksIO._
+
+import StatsIO._
+import SampleCompoundRelationsIO._
+import CompoundAnnotationsIO._
+import RanksIO._
 
 /**
   * This test class tests the whole of the loading and combining functionality. It serves as an example
